refactor upsert repo
